refactor: point location ctor & naming
