+++ packages/phasor/src/elements/brush/brush-element.ts
@@ -12,7 +12,7 @@ import {
   lineIntersects,
   polyLineNearestPoint,
 } from '../../utils/math-utils.js';
-import { PointLocation } from '../../utils/pointLocation.js';
+import { PointLocation } from '../../utils/point-location.js';
 import { type IVec, Vec } from '../../utils/vec.js';
 import { type HitTestOptions, SurfaceElement } from '../surface-element.js';
 import type { IBrush } from './types.js';
