+++ packages/phasor/src/elements/shape/shapes/triangle.ts
@@ -10,7 +10,7 @@ import {
   polygonNearestPoint,
   rotatePoints,
 } from '../../../utils/math-utils.js';
-import { PointLocation } from '../../../utils/pointLocation.js';
+import { PointLocation } from '../../../utils/point-location.js';
 import { type IVec } from '../../../utils/vec.js';
 import type { HitTestOptions } from '../../surface-element.js';
 import type { ShapeElement } from '../shape-element.js';
