+++ packages/phasor/src/elements/shape/types.ts
@@ -1,7 +1,7 @@
 import type { StrokeStyle } from '../../consts.js';
 import type { RoughCanvas } from '../../rough/canvas.js';
 import type { Bound } from '../../utils/bound.js';
-import type { PointLocation } from '../../utils/pointLocation.js';
+import type { PointLocation } from '../../utils/point-location.js';
 import type { IVec } from '../../utils/vec.js';
 import type { HitTestOptions, ISurfaceElement } from '../surface-element.js';
 import type { ShapeElement } from './shape-element.js';
