+++ packages/phasor/src/elements/text/text-element.ts
@@ -6,7 +6,7 @@ import {
   polygonGetPointTangent,
   polygonNearestPoint,
 } from '../../utils/math-utils.js';
-import { PointLocation } from '../../utils/pointLocation.js';
+import { PointLocation } from '../../utils/point-location.js';
 import { type IVec } from '../../utils/vec.js';
 import { SurfaceElement } from '../surface-element.js';
 import type { IText, ITextDelta } from './types.js';
